services/requirements_summarizer.txt
flask
gunicorn
flask_cqlalchemy
redis
numpy
scikit-learn
python-igraph
